+++ go.mod
@@ -2,6 +2,7 @@ module github.com/edgexfoundry/app-functions-sdk-go
 
 require (
 	github.com/BurntSushi/toml v0.3.1
+	github.com/eclipse/paho.mqtt.golang v1.1.1
 	github.com/edgexfoundry/go-mod-core-contracts v0.0.0-20190213225129-211d234d86e4
 	github.com/edgexfoundry/go-mod-registry v0.0.0-20190221211212-171439dc16f8
 	github.com/globalsign/mgo v0.0.0-20181015135952-eeefdecb41b8 // indirect
