+++ pkg/edgexsdk/edgexsdk.go
@@ -28,14 +28,15 @@ import (
 	"syscall"
 
 	"github.com/edgexfoundry/app-functions-sdk-go/internal"
-	"github.com/edgexfoundry/app-functions-sdk-go/pkg/common"
+	"github.com/edgexfoundry/app-functions-sdk-go/internal/common"
+	"github.com/edgexfoundry/app-functions-sdk-go/internal/common/runtime"
+	"github.com/edgexfoundry/app-functions-sdk-go/internal/trigger"
+	"github.com/edgexfoundry/app-functions-sdk-go/internal/trigger/http"
+	"github.com/edgexfoundry/app-functions-sdk-go/internal/trigger/messagebus"
 	"github.com/edgexfoundry/app-functions-sdk-go/pkg/excontext"
-	"github.com/edgexfoundry/app-functions-sdk-go/pkg/runtime"
 	"github.com/edgexfoundry/app-functions-sdk-go/pkg/transforms"
-	"github.com/edgexfoundry/app-functions-sdk-go/pkg/trigger"
-	"github.com/edgexfoundry/app-functions-sdk-go/pkg/trigger/http"
-	"github.com/edgexfoundry/app-functions-sdk-go/pkg/trigger/messagebus"
 	logger "github.com/edgexfoundry/go-mod-core-contracts/clients/logging"
+	"github.com/edgexfoundry/go-mod-core-contracts/models"
 	registry "github.com/edgexfoundry/go-mod-registry"
 	"github.com/edgexfoundry/go-mod-registry/pkg/factory"
 )
@@ -95,6 +96,16 @@ func (sdk *AppFunctionsSDK) HTTPPost(url string) func(excontext.Context, ...inte
 		URL: url,
 	}
 	return transforms.HTTPPost
+}
+
+// MQTTSend ...
+func (sdk *AppFunctionsSDK) MQTTSend(addr models.Addressable, cert string, key string) func(excontext.Context, ...interface{}) (bool, interface{}) {
+	// transforms := transforms.MQTTSender{
+	// 	URL: url,
+	// }
+	sender := transforms.NewMQTTSender(sdk.LoggingClient, addr, cert, key)
+
+	return sender.MQTTSend
 }
 
 //MakeItRun the SDK
