+++ pkg/edgexsdk/edgexsdk_test.go
@@ -20,11 +20,11 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/edgexfoundry/app-functions-sdk-go/pkg/common"
+	"github.com/edgexfoundry/app-functions-sdk-go/internal/common"
+	"github.com/edgexfoundry/app-functions-sdk-go/internal/common/runtime"
+	"github.com/edgexfoundry/app-functions-sdk-go/internal/trigger/http"
+	"github.com/edgexfoundry/app-functions-sdk-go/internal/trigger/messagebus"
 	"github.com/edgexfoundry/app-functions-sdk-go/pkg/excontext"
-	"github.com/edgexfoundry/app-functions-sdk-go/pkg/runtime"
-	"github.com/edgexfoundry/app-functions-sdk-go/pkg/trigger/http"
-	"github.com/edgexfoundry/app-functions-sdk-go/pkg/trigger/messagebus"
 	logger "github.com/edgexfoundry/go-mod-core-contracts/clients/logging"
 	"github.com/stretchr/testify/assert"
 )
