+++ pkg/transforms/http_test.go
@@ -90,6 +90,20 @@ func TestHTTPPostNoParameterPassed(t *testing.T) {
 		t.Fatal("Should have an error when no parameter was passed")
 	}
 }
+func TestHTTPPostInvalidParameter(t *testing.T) {
+	ctx := excontext.Context{
+		LoggingClient: lc,
+	}
+	sender := HTTPSender{}
+	data := "HELLO"
+	continuePipeline, result := sender.HTTPPost(ctx, ([]byte)(data))
+	if continuePipeline != false {
+		t.Fatal("Pipeline should stop")
+	}
+	if result.(error).Error() != "Unexpected type received" {
+		t.Fatal("Should have an error when no parameter was passed")
+	}
+}
 func TestHTTPPostNonExistentEndpoint(t *testing.T) {
 	ctx := excontext.Context{
 		LoggingClient: lc,
